backend/requirements.txt
asgiref==3.2.6
certifi==2019.11.28
chardet==3.0.4
DateTime==4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-oauth-toolkit==1.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
psycopg2==2.8.4
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
zope.interface==5.0.1
urllib3==1.25.8
whitenoise==5.0.1

